refactoring to more accurately follow the flickr file structure and reuse functionality

+++ src/com/company/PhotoAsset.java
@@ -7,13 +7,14 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.MalformedURLException;
 import java.net.URL;
+import java.nio.file.FileAlreadyExistsException;
 
 import static com.company.FlickrAPIEndpoints.getPhotoInfo;
 import static com.company.FlickrAPIEndpoints.getPhotoSizes;
 
 public class PhotoAsset {
+
     private String url;
     private int height;
     private int width;
@@ -32,7 +33,7 @@ public class PhotoAsset {
         this.width = largestPhotoObject.getInt("width");
     }
 
-    public static JSONObject getLargestPhotoSizeObject(JSONObject photoSizeObject){
+    private static JSONObject getLargestPhotoSizeObject(JSONObject photoSizeObject){
         JSONArray sizes = photoSizeObject.getJSONObject("sizes").getJSONArray("size");
         return sizes.getJSONObject(sizes.length()-1);
     }
@@ -45,13 +46,15 @@ public class PhotoAsset {
         InputStream input = new URL(url).openStream();
         String ext = url.substring(url.length()-4);
         File imageFile = new File("media/"+id+"."+ext);
+        if (imageFile.exists() && !imageFile.delete()) {
+            throw new FileAlreadyExistsException(imageFile.getName());
+        }
         if (imageFile.createNewFile()) {
             FileOutputStream outputStream = new FileOutputStream(imageFile);
             outputStream.write(input.readAllBytes());
             outputStream.close();
-            System.out.println("DONE.");
         } else {
-            System.out.println("Error creating file: " + imageFile.getName());
+            throw new IOException("Cannot create file " + imageFile + ". Image not downloaded. ");
         }
         return true;
     }
